refactor(cli): extract skills and mcp commands
